# backend/auth-service/src/main/java/com/notifyhub/auth/api/UpdateProfileRequest.java
package com.notifyhub.auth.api;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record UpdateProfileRequest(
        @NotBlank @Size(max = 80) String firstName,
        @NotBlank @Size(max = 80) String lastName,
        @NotBlank @Size(max = 32) String phoneNumber
) {
}